cleaning up the whole thing. Separated styled components from some jsx sheets, renamed some other sheets
