+++ src/main.jsx
@@ -1,6 +1,5 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import "./index.css";
 import GlobalStyle from "./styles/GlobalStyle.js";
 import Router from "./application/Router.jsx";
