+++ src/styles/GlobalStyle.js
@@ -19,14 +19,11 @@ const GlobalStyle = createGlobalStyle`
 
 }
 
-
- 
  body {
   margin: 0;
   background-color: var(--yellow);
   
  }
-
 
  ul {
   list-style: none;
